+++ broker/src/main/java/pt/ulisboa/tecnico/softeng/broker/services/remote/dataobjects/RestInvoiceData.java
@@ -12,7 +12,7 @@ public class RestInvoiceData {
 	private Long value;
 	@JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
 	private LocalDate date;
-	private Double iva;
+	private Long iva;
 	@JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss.SSS")
 	private DateTime time;
 
@@ -69,11 +69,11 @@ public class RestInvoiceData {
 		this.date = date;
 	}
 
-	public Double getIva() {
+	public Long getIva() {
 		return this.iva;
 	}
 
-	public void setIva(Double iva) {
+	public void setIva(Long iva) {
 		this.iva = iva;
 	}
